[src] Refactored the whole thing, added streaming speech [still needs fixing tho]
