+++ .gitignore
@@ -2,3 +2,4 @@
 .vscode
 baymax_memory.json
 speech.mp3
+test.py
