+++ src/baymax.py
@@ -1,0 +1,24 @@
+
+from speak import *
+from brain import *
+
+#########################################################################################################################
+#########################################################################################################################
+#########################################################################################################################
+#########################################################################################################################
+
+def query_input() -> str:
+    try:
+        return input("> ")
+    except (KeyboardInterrupt, EOFError):
+        print("\n[!] Exiting...")
+        exit(0)
+
+# === MAIN LOOP ===
+if __name__ == "__main__":
+    print("Baymax is here. Say something!")
+    while True:
+        user_input = query_input()
+        reply = ask(user_input)
+        print("Baymax:", reply)
+        speak(reply, True)
